backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
requests==2.31.0
python-decouple==3.8
openai==1.3.0
PyMySQL==1.1.0
cryptography==41.0.7




